refactor(vulkan): centralize dynamic dispatch and move GLSL helpers out of copy.cpp
- Move dtype_to_glsl_storage_type, emit_dynamic_shader_preamble, cast_expr_for_dtype,
  zero_literal_for_dtype, uses_float16/8bit/16bit_storage to shader_compiler.h/cpp
- Move write_descriptor_binding, write_descriptor_buffer, ensure_dynamic_shader_registered,
  is_vulkan_storage_array to kernels.h/cpp
- Add dispatch_dynamic_compute/dispatch_dynamic_compute_begin helpers to kernels.h/cpp
  eliminating ~120 lines of duplicated Vulkan boilerplate across 3 dispatch functions
- Refactor dispatch_dynamic_offset_kernel, dispatch_dynamic_general_copy, and
  dispatch_dynamic_vector_cast_copy to use dispatch_dynamic_compute
- Convert vector cast shader to use push constants for offsets instead of baking
  them into GLSL, reducing shader variants from O(layouts) to O(dtype pairs)
